CNDB-14144: Change Version.LATEST to ED. Move testIndexMetaForNumRows from BM25Test to FeaturesVersionSupportTest and extend coverage for all versions. Add FeaturesVersionSupportEDTest
