+++ src/java/org/apache/cassandra/index/sai/disk/format/Version.java
@@ -74,7 +74,7 @@ public class Version implements Comparable<Version>
     public static final Version VECTOR_EARLIEST = BA;
     public static final Version JVECTOR_EARLIEST = CA;
     public static final Version BM25_EARLIEST = EC;
-    public static final Version LATEST = EC;
+    public static final Version LATEST = ED;
     // The current version can be configured to be an earlier version to support partial upgrades that don't
     // write newer versions of the on-disk formats. This is volatile rather than final so that tests may
     // use reflection to change it and safely publish across threads.
