+++ test/unit/org/apache/cassandra/index/sai/SAITester.java
@@ -28,6 +28,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Random;
@@ -36,6 +38,7 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import javax.management.AttributeNotFoundException;
@@ -237,6 +240,68 @@ public class SAITester extends CQLTester
     {
         CQLTester.setUpClass();
         CQLTester.enableCoordinatorExecution();
+    }
+
+    // Pattern that treats apostrophes within words as part of the word
+    protected static final Pattern PATTERN = Pattern.compile("[^\\w']+|'(?=\\s)|(?<=\\s)'");
+    protected static final int DATASET_BODY_COLUMN = 3;
+
+    protected final static Object[][] DATASET =
+    {
+    { 1, "Climate", 5, "Climate change is a pressing issue. Climate patterns are shifting globally. Scientists study climate data daily.", 1 },
+    { 2, "Technology", 3, "Technology is advancing. New technology in AI and robotics is groundbreaking.", 1 },
+    { 3, "Economy", 4, "The economy is recovering. Economy experts are optimistic. However, the global economy still faces risks.", 1 },
+    { 4, "Health", 3, "Health is wealth. Health policies need to be improved to ensure better public health outcomes.", 1 },
+    { 5, "Education", 2, "Education is the foundation of success. Online education is booming.", 4 },
+    { 6, "Climate", 4, "Climate and health are closely linked. Climate affects air quality and health outcomes.", 2 },
+    { 7, "Education", 3, "Technology and education go hand in hand. EdTech is revolutionizing education through technology.", 3 },
+    { 8, "Economy", 3, "The global economy is influenced by technology. Fintech is a key part of the economy today.", 2 },
+    { 9, "Health", 3, "Education and health programs must be prioritized. Health education is vital in schools.", 2 },
+    { 10, "Mixed", 3, "Technology, economy, and education are pillars of development.", 2 },
+    { 11, "Climate", 5, "Climate climate climate. It's everywhere. Climate drives political and economic decisions.", 1 },
+    { 12, "Health", 2, "Health concerns rise with climate issues. Health organizations are sounding the alarm.", 2 },
+    { 13, "Economy", 3, "The economy is fluctuating. Uncertainty looms over the economy.", 1 },
+    { 14, "Health", 3, "Cutting-edge technology is transforming healthcare. Healthtech merges health and technology.", 1 },
+    { 15, "Education", 2, "Education reforms are underway. Education experts suggest holistic changes.", 1 },
+    { 16, "Climate", 4, "Climate affects the economy and health. Climate events cost billions annually.", 1 },
+    { 17, "Technology", 3, "Technology is the backbone of the modern economy. Without technology, economic growth stagnates.", 2 },
+    { 18, "Health", 2, "Health is discussed less than economy or climate or technology, but health matters deeply.", 1 },
+    { 19, "Climate", 5, "Climate change, climate policies, climate research—climate is the buzzword of our time.", 2 },
+    { 20, "Mixed", 3, "Investments in education and technology will shape the future of the global economy.", 1 }
+    };
+
+    protected void insertCollectionData()
+    {
+        int setsize = 1;
+        for (int row = 0; row < DATASET.length; row++)
+        {
+            var set = new HashSet<String>();
+            for (int j = 0; j < setsize; j++)
+                set.add((String) DATASET[row - j][3]);
+            if (setsize >= 3)
+                setsize -= 2;
+            else
+                setsize++;
+            var map = new HashMap<Integer, String>();
+            var map_text = new HashMap<String, String>();
+            for (int j = 0; j <= row && j < 3; j++)
+            {
+                map.putIfAbsent((Integer) DATASET[row - j][2], (String) DATASET[row - j][1]);
+                map_text.putIfAbsent((String) DATASET[row - j][1], (String) DATASET[row - j][3]);
+            }
+
+            execute(
+            "INSERT INTO %s (id, category, score, body, bodyset, map_category, map_body) " +
+            "VALUES (?, ?, ?, ?, ?, ?, ?)",
+            DATASET[row][0],
+            DATASET[row][1],
+            DATASET[row][2],
+            DATASET[row][3],
+            set,
+            map,
+            map_text
+            );
+        }
     }
 
     /**
