+++ test/unit/org/apache/cassandra/index/sai/cql/BM25Test.java
@@ -18,23 +18,14 @@ package org.apache.cassandra.index.sai.cql;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
-import org.apache.cassandra.index.sai.SSTableIndex;
-import org.apache.cassandra.index.sai.memory.MemtableIndex;
-import org.apache.cassandra.index.sai.memory.TrieMemoryIndex;
-import org.apache.cassandra.index.sai.memory.TrieMemtableIndex;
 import org.assertj.core.api.Assertions;
 
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -66,10 +57,6 @@ public class BM25Test extends SAITester
     {
         return Arrays.asList(new Object[]{ Version.BM25_EARLIEST, Version.ED });
     }
-
-    // Pattern that treats apostrophes within words as part of the word
-    private static final Pattern PATTERN = Pattern.compile("[^\\w']+|'(?=\\s)|(?<=\\s)'");
-    public static final int DATASET_BODY_COLUMN = 3;
 
     @Before
     public void setup() throws Throwable
@@ -842,115 +829,6 @@ public class BM25Test extends SAITester
                 "climate");
     }
 
-    /**
-     * Asserts that memtable SAI index maintains expected row count, which is, then,
-     * used to store row count in SSTable SAI index and its segments. This is also
-     * asserted.
-     */
-    @Test
-    public void testIndexMetaForNumRows()
-    {
-        SAIUtil.setCurrentVersion(Version.ED);
-
-        createTable("CREATE TABLE %s (id int PRIMARY KEY, category text, score int, " +
-                    "title text, body text, bodyset set<text>, " +
-                    "map_category map<int, text>, map_body map<text, text>)");
-        String bodyIndexName = createAnalyzedIndex("body", true);
-        String scoreIndexName = createIndex("CREATE CUSTOM INDEX ON %s (score) USING 'StorageAttachedIndex'");
-        String mapIndexName = createIndex("CREATE CUSTOM INDEX ON %s (map_category) USING 'StorageAttachedIndex'");
-        insertCollectionData();
-        int totalTermsCount = IntStream.range(0, DATASET.length)
-                                       .map(this::calculateTotalTermsForRow)
-                                       .sum();
-
-        assertNumRowsMemtable(scoreIndexName, DATASET.length);
-        assertNumRowsMemtable(bodyIndexName, DATASET.length);
-        assertNumRowsMemtable(mapIndexName, DATASET.length);
-        execute("DELETE FROM %s WHERE id = ?", 5);
-        totalTermsCount -= calculateTotalTermsForRow(4);
-        flush();
-        assertNumRowsAndTotalTermsSSTable(scoreIndexName, DATASET.length - 1, DATASET.length - 1);
-        assertNumRowsAndTotalTermsSSTable(bodyIndexName, DATASET.length - 1, totalTermsCount);
-        assertNumRowsSSTable(mapIndexName, DATASET.length - 1);
-        execute("DELETE FROM %s WHERE id = ?", 10);
-        flush();
-        assertNumRowsAndTotalTermsSSTable(scoreIndexName, DATASET.length - 1, DATASET.length - 1);
-        assertNumRowsAndTotalTermsSSTable(bodyIndexName, DATASET.length - 1, totalTermsCount);
-        assertNumRowsSSTable(mapIndexName, DATASET.length - 1);
-        compact();
-        totalTermsCount -= calculateTotalTermsForRow(9);
-        assertNumRowsAndTotalTermsSSTable(scoreIndexName, DATASET.length - 2, DATASET.length - 2);
-        assertNumRowsAndTotalTermsSSTable(bodyIndexName, DATASET.length - 2, totalTermsCount);
-        assertNumRowsSSTable(mapIndexName, DATASET.length - 2);
-    }
-
-    private void assertNumRowsMemtable(String indexName, int expectedNumRows)
-    {
-        int rowCount = 0;
-
-        for (var memtable : getCurrentColumnFamilyStore().getAllMemtables())
-        {
-            MemtableIndex memIndex = getIndexContext(indexName).getLiveMemtables().get(memtable);
-            assert memIndex instanceof TrieMemtableIndex;
-            rowCount = Arrays.stream(((TrieMemtableIndex) memIndex).getRangeIndexes())
-                             .map(index -> ((TrieMemoryIndex) index).getDocLengths().size())
-                             .mapToInt(Integer::intValue).sum();
-        }
-        assertEquals(expectedNumRows, rowCount);
-    }
-
-    private void assertNumRowsSSTable(String indexName, int expectedNumRows)
-    {
-        assertNumRowsAndTotalTermsSSTable(indexName, expectedNumRows, -1);
-    }
-
-    private void assertNumRowsAndTotalTermsSSTable(String indexName, int expectedNumRows, int expectedTotalTermsCount
-    )
-    {
-        long indexRowCount = 0;
-        long segmentRowCount = 0;
-        long totalTermCount = 0;
-        for (SSTableIndex sstableIndex : getIndexContext(indexName).getView())
-        {
-            indexRowCount += sstableIndex.getRowCount();
-            for (var segment : sstableIndex.getSegments())
-            {
-                var metadata = segment.metadata;
-                Assert.assertNotNull(metadata);
-                segmentRowCount += metadata.numRows;
-                totalTermCount += metadata.totalTermCount;
-            }
-        }
-        assertEquals(indexRowCount, segmentRowCount);
-        assertEquals(expectedNumRows, indexRowCount);
-        if (expectedTotalTermsCount > 0)
-            assertEquals(expectedTotalTermsCount, totalTermCount);
-    }
-
-    private final static Object[][] DATASET =
-    {
-    { 1, "Climate", 5, "Climate change is a pressing issue. Climate patterns are shifting globally. Scientists study climate data daily.", 1 },
-    { 2, "Technology", 3, "Technology is advancing. New technology in AI and robotics is groundbreaking.", 1 },
-    { 3, "Economy", 4, "The economy is recovering. Economy experts are optimistic. However, the global economy still faces risks.", 1 },
-    { 4, "Health", 3, "Health is wealth. Health policies need to be improved to ensure better public health outcomes.", 1 },
-    { 5, "Education", 2, "Education is the foundation of success. Online education is booming.", 4 },
-    { 6, "Climate", 4, "Climate and health are closely linked. Climate affects air quality and health outcomes.", 2 },
-    { 7, "Education", 3, "Technology and education go hand in hand. EdTech is revolutionizing education through technology.", 3 },
-    { 8, "Economy", 3, "The global economy is influenced by technology. Fintech is a key part of the economy today.", 2 },
-    { 9, "Health", 3, "Education and health programs must be prioritized. Health education is vital in schools.", 2 },
-    { 10, "Mixed", 3, "Technology, economy, and education are pillars of development.", 2 },
-    { 11, "Climate", 5, "Climate climate climate. It's everywhere. Climate drives political and economic decisions.", 1 },
-    { 12, "Health", 2, "Health concerns rise with climate issues. Health organizations are sounding the alarm.", 2 },
-    { 13, "Economy", 3, "The economy is fluctuating. Uncertainty looms over the economy.", 1 },
-    { 14, "Health", 3, "Cutting-edge technology is transforming healthcare. Healthtech merges health and technology.", 1 },
-    { 15, "Education", 2, "Education reforms are underway. Education experts suggest holistic changes.", 1 },
-    { 16, "Climate", 4, "Climate affects the economy and health. Climate events cost billions annually.", 1 },
-    { 17, "Technology", 3, "Technology is the backbone of the modern economy. Without technology, economic growth stagnates.", 2 },
-    { 18, "Health", 2, "Health is discussed less than economy or climate or technology, but health matters deeply.", 1 },
-    { 19, "Climate", 5, "Climate change, climate policies, climate research—climate is the buzzword of our time.", 2 },
-    { 20, "Mixed", 3, "Investments in education and technology will shape the future of the global economy.", 1 }
-    };
-
     private void analyzeDataset(String term)
     {
         for (Object[] row : DATASET)
@@ -969,12 +847,6 @@ public class BM25Test extends SAITester
         }
     }
 
-    private int calculateTotalTermsForRow(int row)
-    {
-        String body = (String) DATASET[row][DATASET_BODY_COLUMN];
-        return PATTERN.split(body.toLowerCase()).length;
-    }
-
     private void insertPrimitiveData()
     {
         insertPrimitiveData(0, DATASET.length);
@@ -991,40 +863,6 @@ public class BM25Test extends SAITester
             row[1],
             row[2],
             row[3]
-            );
-        }
-    }
-
-    private void insertCollectionData()
-    {
-        int setsize = 1;
-        for (int row = 0; row < DATASET.length; row++)
-        {
-            var set = new HashSet<String>();
-            for (int j = 0; j < setsize; j++)
-                set.add((String) DATASET[row - j][3]);
-            if (setsize >= 3)
-                setsize -= 2;
-            else
-                setsize++;
-            var map = new HashMap<Integer, String>();
-            var map_text = new HashMap<String, String>();
-            for (int j = 0; j <= row && j < 3; j++)
-            {
-                map.putIfAbsent((Integer) DATASET[row - j][2], (String) DATASET[row - j][1]);
-                map_text.putIfAbsent((String) DATASET[row - j][1], (String) DATASET[row - j][3]);
-            }
-
-            execute(
-            "INSERT INTO %s (id, category, score, body, bodyset, map_category, map_body) " +
-            "VALUES (?, ?, ?, ?, ?, ?, ?)",
-            DATASET[row][0],
-            DATASET[row][1],
-            DATASET[row][2],
-            DATASET[row][3],
-            set,
-            map,
-            map_text
             );
         }
     }
